Cleanup plugins, fix pathPrefix logic bug
