+++ package.json
@@ -41,6 +41,7 @@
     "@leafygreen-ui/text-input": "^5.0.10",
     "@reach/router": "^1.3.4",
     "@testing-library/react-hooks": "^5.1.1",
+    "axios": "^0.21.1",
     "copy-to-clipboard": "^3.3.1",
     "css-loader": "^5.2.1",
     "dlv": "^1.1.3",
@@ -48,6 +49,7 @@
     "esm": "^3.2.25",
     "fast-xml-parser": "^3.19.0",
     "gatsby": "^2.31.1",
+    "gatsby-node-helpers": "^0.3.0",
     "gatsby-plugin-alias-imports": "^1.0.5",
     "gatsby-plugin-emotion": "^5.0.0",
     "gatsby-plugin-feed": "^2.12.0",
@@ -57,9 +59,12 @@
     "gatsby-plugin-react-helmet": "^3.9.0",
     "gatsby-plugin-sitemap": "^2.12.0",
     "gatsby-plugin-webpack-bundle-analyser-v2": "^1.1.21",
+    "gatsby-source-filesystem": "^3.4.0",
+    "lodash": "^4.17.21",
     "memoizerific": "^1.11.3",
     "mongodb-stitch-browser-sdk": "^4.8.0",
     "mongodb-stitch-server-sdk": "^4.8.0",
+    "pluralize": "^8.0.0",
     "postcss": "^8.2.10",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
