--- plugins/gatsby-new-source-strapi/.babelrc
@@ -1,14 +1,0 @@
-{
-  "sourceMaps": true,
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "node": 10
-        }
-      }
-    ]
-  ],
-  "plugins": ["@babel/plugin-transform-runtime"]
-}
