--- plugins/gatsby-new-source-strapi/.eslintcache
@@ -1,1 +1,0 @@
-[{"/Users/alex/dev/strapi/expansion/gatsby-source-strapi/src/index.js":"1","/Users/alex/dev/strapi/expansion/gatsby-source-strapi/src/normalize.js":"2"},{"size":2844,"mtime":1617718155964,"results":"3","hashOfConfig":"4"},{"size":1950,"mtime":1611916768939,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"zn7rtp",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/alex/dev/strapi/expansion/gatsby-source-strapi/src/index.js",[],"/Users/alex/dev/strapi/expansion/gatsby-source-strapi/src/normalize.js",[]]
