--- plugins/gatsby-new-source-strapi/.eslintrc.js
@@ -1,12 +1,0 @@
-module.exports = {
-  env: {
-    es2020: true,
-    node: true,
-  },
-  extends: 'eslint:recommended',
-  parserOptions: {
-    ecmaVersion: 11,
-    sourceType: 'module',
-  },
-  rules: {},
-};
