--- plugins/gatsby-new-source-strapi/.prettierrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  endOfLine: 'lf',
-  semi: true,
-  singleQuote: true,
-  tabWidth: 2,
-  trailingComma: 'es5',
-  printWidth: 100,
-};
