+++ plugins/gatsby-source-strapi/gatsby-node.js
