+++ plugins/gatsby-source-strapi/package.json
